requirements.txt
pandas
scikit-learn
torchvision
numpy<2.0
matplotlib
reformer_pytorch
